# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Describe It - Spanish Learning API
  version: 1.0.0
  description: |
    A comprehensive API for visual-based Spanish learning through AI-powered image descriptions, 
    Q&A generation, vocabulary extraction, and translation services.
    
    This API enables users to:
    - Generate multi-style image descriptions in Spanish and English
    - Create interactive Q&A sessions for comprehension testing
    - Extract and translate vocabulary from descriptions
    - Track learning progress and export data
    - Search and manage images for learning content
    
  contact:
    name: Describe It Support
    email: support@describe-it.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://describe-it.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /settings/save:
    post:
      tags:
        - Settings
      summary: Save User Settings
      description: Save user preferences and settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettings'
      responses:
        '200':
          description: Settings saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserSettings:
      type: object
      properties:
        preferredLanguage:
          type: string
          enum: ["es", "en"]
          default: "es"
          description: User's preferred language for learning
          example: "es"
        difficultyLevel:
          type: string
          enum: ["beginner", "intermediate", "advanced"]
          default: "beginner"
          description: User's learning level
          example: "beginner"
        autoTranslate:
          type: boolean
          default: true
          description: Whether to auto-translate vocabulary
          example: true
        showHints:
          type: boolean
          default: true
          description: Whether to show learning hints
          example: true
...
    SettingsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        settings:
          $ref: '#/components/schemas/UserSettings'
...